requirements.txt
Flask==2.0.3
Pillow==9.0.1